Consolidate hub readiness polling and grant issuance into the harness
The hub-subprocess harness and the dev orchestrator each carried a
private copy of the same readiness poll, and the hub-subprocess
harness inlined the same grant statements the shared grant helper
issues. Both now consume the test-harness module, leaving one
implementation of each on the bun side; the browser harness keeps its
own poll because its runtime cannot resolve workspace source imports.
